move fraudproof and receipts api to SettlementApi
