+++ crates/sc-consensus-fraud-proof/Cargo.toml
@@ -18,4 +18,5 @@ sp-api = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate",
 sp-consensus = { version = "0.10.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
 sp-domains = { version = "0.1.0", path = "../sp-domains" }
 sp-runtime = { version = "7.0.0", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
+sp-settlement = { version = "0.1.0", path = "../sp-settlement" }
 subspace-fraud-proof = { version = "0.1.0", path = "../subspace-fraud-proof" }
