+++ crates/sc-consensus-fraud-proof/src/lib.rs
@@ -20,8 +20,9 @@ use codec::{Decode, Encode};
 use sc_consensus::block_import::{BlockCheckParams, BlockImport, BlockImportParams, ImportResult};
 use sp_api::{ProvideRuntimeApi, TransactionFor};
 use sp_consensus::Error as ConsensusError;
-use sp_domains::{DomainId, ExecutorApi};
+use sp_domains::DomainId;
 use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
+use sp_settlement::SettlementApi;
 use std::marker::PhantomData;
 use std::sync::Arc;
 use subspace_fraud_proof::VerifyFraudProof;
@@ -62,7 +63,7 @@ impl<Block, Client, Inner, Verifier, DomainHash> BlockImport<Block>
 where
     Block: BlockT,
     Client: ProvideRuntimeApi<Block> + Send + Sync + 'static,
-    Client::Api: ExecutorApi<Block, DomainHash>,
+    Client::Api: SettlementApi<Block, DomainHash>,
     Inner: BlockImport<Block, Transaction = TransactionFor<Client, Block>, Error = ConsensusError>
         + Send,
     Verifier: VerifyFraudProof<Block> + Send,
