+++ crates/sp-settlement/src/lib.rs
@@ -19,7 +19,8 @@
 
 use codec::{Decode, Encode};
 use sp_core::H256;
-use sp_domains::DomainId;
+use sp_domains::fraud_proof::FraudProof;
+use sp_domains::{DomainId, ExecutionReceipt};
 use sp_runtime::traits::NumberFor;
 use sp_std::vec::Vec;
 
@@ -52,5 +53,20 @@ sp_api::decl_runtime_apis! {
 
         /// Returns the maximum receipt drift.
         fn maximum_receipt_drift() -> NumberFor<Block>;
+
+        /// Extract the receipts from the given extrinsics.
+        fn extract_receipts(
+            extrinsics: Vec<Block::Extrinsic>,
+            domain_id: DomainId,
+        ) -> Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>;
+
+        /// Extract the fraud proofs from the given extrinsics.
+        fn extract_fraud_proofs(
+            extrinsics: Vec<Block::Extrinsic>,
+            domain_id: DomainId,
+        ) -> Vec<FraudProof<NumberFor<Block>, Block::Hash>>;
+
+        /// Submits the fraud proof via an unsigned extrinsic.
+        fn submit_fraud_proof_unsigned(fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>);
     }
 }
